+++ src/main/java/nom/bdezonia/zorbage/type/complex/float16/ComplexFloat16VectorMember.java
@@ -737,8 +737,8 @@ public final class ComplexFloat16VectorMember
 		if (component > 1) return BigInteger.ZERO;
 		ComplexFloat16Member tmp = tmpComp.get();
 		getV(index.get(0), tmp);
-		if (component == 0) return BigDecimal.valueOf(tmp.r()).toBigInteger();
-		return BigDecimal.valueOf(tmp.i()).toBigInteger();
+		if (component == 0) return BigInteger.valueOf((long) tmp.r());
+		return BigInteger.valueOf((long) tmp.i());
 	}
 
 	@Override
@@ -965,8 +965,8 @@ public final class ComplexFloat16VectorMember
 		else {
 			ComplexFloat16Member tmp = tmpComp.get();
 			getV(index.get(0), tmp);
-			if (component == 0) return BigDecimal.valueOf(tmp.r()).toBigInteger();
-			return BigDecimal.valueOf(tmp.i()).toBigInteger();
+			if (component == 0) return BigInteger.valueOf((long) tmp.r());
+			return BigInteger.valueOf((long) tmp.i());
 		}
 	}
 
@@ -1264,8 +1264,8 @@ public final class ComplexFloat16VectorMember
 		BigInteger[] values = new BigInteger[2 * (int) storage.size()];
 		for (int i = 0, k = 0; i < storage.size(); i++) {
 			storage.get(i, value);
-			values[k++] = BigDecimal.valueOf(value.r()).toBigInteger();
-			values[k++] = BigDecimal.valueOf(value.i()).toBigInteger();
+			values[k++] = BigInteger.valueOf((long) value.r());
+			values[k++] = BigInteger.valueOf((long) value.i());
 		}
 		return values;
 	}
