+++ src/main/java/nom/bdezonia/zorbage/type/quaternion/float16/QuaternionFloat16CartesianTensorProductMember.java
@@ -966,15 +966,15 @@ public final class QuaternionFloat16CartesianTensorProductMember
 		getV(index, tmp);
 		if (component < 2) {
 			if (component == 0)
-				return BigDecimal.valueOf(tmp.r()).toBigInteger();
+				return BigInteger.valueOf((long) tmp.r());
 			else
-				return BigDecimal.valueOf(tmp.i()).toBigInteger();
+				return BigInteger.valueOf((long) tmp.i());
 		}
 		else { // component >= 2
 			if (component == 2)
-				return BigDecimal.valueOf(tmp.j()).toBigInteger();
+				return BigInteger.valueOf((long) tmp.j());
 			else
-				return BigDecimal.valueOf(tmp.k()).toBigInteger();
+				return BigInteger.valueOf((long) tmp.k());
 		}
 	}
 
@@ -1147,15 +1147,15 @@ public final class QuaternionFloat16CartesianTensorProductMember
 			getV(index, tmp);
 			if (component < 2) {
 				if (component == 0)
-					return BigDecimal.valueOf(tmp.r()).toBigInteger();
+					return BigInteger.valueOf((long) tmp.r());
 				else
-					return BigDecimal.valueOf(tmp.i()).toBigInteger();
+					return BigInteger.valueOf((long) tmp.i());
 			}
 			else { // component >= 2
 				if (component == 2)
-					return BigDecimal.valueOf(tmp.j()).toBigInteger();
+					return BigInteger.valueOf((long) tmp.j());
 				else
-					return BigDecimal.valueOf(tmp.k()).toBigInteger();
+					return BigInteger.valueOf((long) tmp.k());
 			}
 		}
 	}
@@ -1483,10 +1483,10 @@ public final class QuaternionFloat16CartesianTensorProductMember
 		BigInteger[] values = new BigInteger[4 * (int) storage.size()];
 		for (int i = 0, k = 0; i < storage.size(); i++) {
 			storage.get(i, value);
-			values[k++] = BigDecimal.valueOf(value.r()).toBigInteger();
-			values[k++] = BigDecimal.valueOf(value.i()).toBigInteger();
-			values[k++] = BigDecimal.valueOf(value.j()).toBigInteger();
-			values[k++] = BigDecimal.valueOf(value.k()).toBigInteger();
+			values[k++] = BigInteger.valueOf((long) value.r());
+			values[k++] = BigInteger.valueOf((long) value.i());
+			values[k++] = BigInteger.valueOf((long) value.j());
+			values[k++] = BigInteger.valueOf((long) value.k());
 		}
 		return values;
 	}
