+++ src/main/java/nom/bdezonia/zorbage/type/quaternion/float16/QuaternionFloat16MatrixMember.java
@@ -1072,16 +1072,16 @@ public final class QuaternionFloat16MatrixMember
 		if (component < 2) {
 			// 0 <= component <= 1
 			if (component == 0)
-				return BigDecimal.valueOf(tmp.r()).toBigInteger();
+				return BigInteger.valueOf((long) tmp.r());
 			else // component == 1
-				return BigDecimal.valueOf(tmp.i()).toBigInteger();
+				return BigInteger.valueOf((long) tmp.i());
 		}
 		else {
 			// 2 <= component <= 3
 			if (component == 2)
-				return BigDecimal.valueOf(tmp.j()).toBigInteger();
+				return BigInteger.valueOf((long) tmp.j());
 			else // component == 3
-				return BigDecimal.valueOf(tmp.k()).toBigInteger();
+				return BigInteger.valueOf((long) tmp.k());
 		}
 	}
 
@@ -1438,16 +1438,16 @@ public final class QuaternionFloat16MatrixMember
 			if (component < 2) {
 				// 0 <= component <= 1
 				if (component == 0)
-					return BigDecimal.valueOf(tmp.r()).toBigInteger();
+					return BigInteger.valueOf((long) tmp.r());
 				else // component == 1
-					return BigDecimal.valueOf(tmp.i()).toBigInteger();
+					return BigInteger.valueOf((long) tmp.i());
 			}
 			else {
 				// 2 <= component <= 3
 				if (component == 2)
-					return BigDecimal.valueOf(tmp.j()).toBigInteger();
+					return BigInteger.valueOf((long) tmp.j());
 				else // component == 3
-					return BigDecimal.valueOf(tmp.k()).toBigInteger();
+					return BigInteger.valueOf((long) tmp.k());
 			}
 		}
 	}
@@ -1803,10 +1803,10 @@ public final class QuaternionFloat16MatrixMember
 		BigInteger[] values = new BigInteger[4 * (int) storage.size()];
 		for (int i = 0, k = 0; i < storage.size(); i++) {
 			storage.get(i, value);
-			values[k++] = BigDecimal.valueOf(value.r()).toBigInteger();
-			values[k++] = BigDecimal.valueOf(value.i()).toBigInteger();
-			values[k++] = BigDecimal.valueOf(value.j()).toBigInteger();
-			values[k++] = BigDecimal.valueOf(value.k()).toBigInteger();
+			values[k++] = BigInteger.valueOf((long) value.r());
+			values[k++] = BigInteger.valueOf((long) value.i());
+			values[k++] = BigInteger.valueOf((long) value.j());
+			values[k++] = BigInteger.valueOf((long) value.k());
 		}
 		return values;
 	}
